Assignment 9: Final submission file structure ready.
